Max scale added to smart resize
Stage width/height will be kept around the given target values up to a
point where view port scale exceeds the maximum given scale - after
that stage width and height will be increased.

This allows for keeping the same stage width/height values for i.e.
iPhone 3Gs and iPhone 4, but will make stage a bit wider and higher
when the same app is launched on an iPad.